+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## Current
 - fixed issue in After hook if browser is not started
+- added I save css property step
+- added I expect css property step
 - added capability to connect to playwright server